usb: gadget: uvc: Fix compile error due to type mismatch

+++ drivers/usb/gadget/function/uvc_video.c
@@ -306,7 +306,7 @@ static int uvcg_video_usb_req_queue(struct uvc_video *video,
 		 */
 		if (list_empty(&video->req_free) || ureq->last_buf ||
 			!(video->req_int_count %
-			min(DIV_ROUND_UP(video->uvc_num_requests, 4), UVCG_REQ_MAX_INT_COUNT))) {
+			min(DIV_ROUND_UP(video->uvc_num_requests, 4), (unsigned int)UVCG_REQ_MAX_INT_COUNT))) {
 			video->req_int_count = 0;
 			req->no_interrupt = 0;
 		} else {
